Refactoring: compliance of methods, model fields validated, fix requests

+++ src/wishes/wishes.controller.ts
@@ -1,13 +1,17 @@
 import {
+  UnsupportedMediaTypeException,
+  NotFoundException,
   ParseIntPipe,
   Controller,
+  HttpStatus,
   UseGuards,
+  HttpCode,
   Delete,
   Param,
-  Body,
-  Get,
   Patch,
+  Body,
   Post,
+  Get,
   Req,
 } from '@nestjs/common';
 import { CreateWishDto } from './dto/create-wish.dto';
@@ -19,10 +23,17 @@ import { JwtGuard } from '../guards/jwt.guard';
 export class WishesController {
   constructor(private readonly wishesService: WishesService) {}
 
-  @Post()
   @UseGuards(JwtGuard)
-  create(@Body() createWishDto: CreateWishDto, @Req() req) {
-    return this.wishesService.create(createWishDto, req.user.id);
+  @Post()
+  @HttpCode(HttpStatus.CREATED)
+  async create(@Body() createWishDto: CreateWishDto, @Req() req) {
+    const acceptHeader = req.headers['accept'];
+    if (!acceptHeader || !acceptHeader.includes('application/json')) {
+      throw new UnsupportedMediaTypeException('Unsupported Media Type');
+    }
+
+    await this.wishesService.create(createWishDto, req.user.id);
+    return {};
   }
 
   @Get('last')
@@ -35,34 +46,63 @@ export class WishesController {
     return this.wishesService.findTopWish();
   }
 
-  @Get(':id')
   @UseGuards(JwtGuard)
-  getOneWish(@Param('id', ParseIntPipe) id: number) {
-    return this.wishesService.findOneWish({
+  @Get(':id')
+  async getOneWish(@Param('id', ParseIntPipe) id: number) {
+    const wish = await this.wishesService.findOneWish({
       where: { id: id },
       relations: { owner: true, offers: true },
     });
+
+    if (!wish) {
+      throw new NotFoundException('Wish not found');
+    }
+
+    return wish;
   }
 
-  @Patch(':id')
   @UseGuards(JwtGuard)
-  update(
+  @Patch(':id')
+  async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateWishDto: UpdateWishDto,
     @Req() req,
   ) {
-    return this.wishesService.update(id, updateWishDto, req.user.id);
+    const updatedWish = await this.wishesService.update(
+      id,
+      updateWishDto,
+      req.user.id,
+    );
+
+    if (!updatedWish) {
+      throw new NotFoundException('Wish not found');
+    }
+
+    return {};
   }
 
-  @Post(':id/copy')
   @UseGuards(JwtGuard)
-  copy(@Param('id') id: number, @Req() req) {
-    return this.wishesService.copy(id, req.user.id);
-  }
-
   @Delete(':id')
+  @HttpCode(HttpStatus.OK)
+  async remove(@Param('id', ParseIntPipe) id: number, @Req() req) {
+    const acceptHeader = req.headers['accept'];
+    if (!acceptHeader || !acceptHeader.includes('application/json')) {
+      throw new UnsupportedMediaTypeException('Unsupported Media Type');
+    }
+
+    await this.wishesService.remove(id, req.user.id);
+  }
+
   @UseGuards(JwtGuard)
-  remove(@Param('id') id: number, @Req() req) {
-    return this.wishesService.remove(id, req.user.id);
+  @Post(':id/copy')
+  @HttpCode(HttpStatus.CREATED)
+  async copy(@Param('id') id: number, @Req() req) {
+    const acceptHeader = req.headers['accept'];
+    if (!acceptHeader || !acceptHeader.includes('application/json')) {
+      throw new UnsupportedMediaTypeException('Unsupported Media Type');
+    }
+
+    await this.wishesService.copy(id, req.user.id);
+    return {};
   }
 }
